Upgrade to Android 33
This upgrades the target project from Android 31 (12) to Android 33 (13). Some additional permissions are needed to ensure the app is allowed to open in the background. Some of these may not be needed, but the changes are currently being tested.

So far, things look good though. The app is able to respawn and connect to N33ble after a 3 hour background delay (where the app has been killed).