Fix error on gatt disconnect and improve gatt retry logic
N33ble1BluetoothGattCallback was frequently (80% of the time) failing with a BleServiceError when N33ble1 was disconnecting/disappearing, perhaps introduced by an Android update.

N33ble1BluetoothGattCallback now notifies N33ble1MonitorService when it encounters an error, and the monitor service now allows retries (up to 3x within 30 seconds, using EventThrottler) by disconnecting and trying to reconnect to N33ble1. This will effectively cleanly disconnect from N33ble1 when it disappears instead of erroring, not reconnecting, and not removing the foreground notification.